+++ R/theme_gtsummary_roche.R
@@ -8,6 +8,10 @@
 #' - Font size defaults are 8 points for all the table by the footers that are 7 points.
 #' - Border defaults to `flextable::fp_border_default(width = 0.5)`.
 #' - The `add_overall(col_label)` default value has been updated.
+#' - The results from `gtsummary::tbl_hierachrical()` and `gtsummary::tbl_hierachrical_count()`
+#'   are now post-processed with `gtsummary::remove_footnote_header()`,
+#'   `crane::modify_zero_recode()`, and `crane::modify_header_rm_md()`.
+#'
 #'
 #' @inheritParams gtsummary::theme_gtsummary_printer
 #' @inheritParams gtsummary::theme_gtsummary_compact
@@ -36,24 +40,7 @@ theme_gtsummary_roche <- function(font_size = NULL,
   border <- flextable::fp_border_default(width = 0.5)
 
   # Initialization with compact gt options -------------------------------------
-  lst_theme <-
-    list(
-      "pkgwide-str:theme_name" = "Roche",
-      # compact gt tables
-      "as_gt-lst:addl_cmds" = list(
-        tab_spanner = rlang::expr(
-          gt::tab_options(
-            table.font.size = !!(font_size %||% 13),
-            data_row.padding = gt::px(1),
-            summary_row.padding = gt::px(1),
-            grand_summary_row.padding = gt::px(1),
-            footnotes.padding = gt::px(1),
-            source_notes.padding = gt::px(1),
-            row_group.padding = gt::px(1)
-          )
-        )
-      )
-    )
+  lst_theme <- list("pkgwide-str:theme_name" = "Roche")
 
   # updating with some pharma-specific bits ------------------------------------
   lst_theme <- lst_theme |>
@@ -62,7 +49,19 @@ theme_gtsummary_roche <- function(font_size = NULL,
         "tbl_summary-fn:percent_fun" = label_roche_percent(),
         "pkgwide-fn:pvalue_fun" = label_roche_pvalue(),
         "add_overall.tbl_summary-arg:col_label" = "All Participants  \nN = {gtsummary::style_number(N)}",
-        "pkgwide-str:print_engine" = print_engine
+        "pkgwide-str:print_engine" = print_engine,
+        "tbl_hierarchical-fn:addnl-fn-to-run" =
+          \(x) {
+            gtsummary::remove_footnote_header(x) |>
+              modify_zero_recode() |>
+              modify_header_rm_md()
+          },
+        "tbl_hierarchical_count-fn:addnl-fn-to-run" =
+          \(x) {
+            gtsummary::remove_footnote_header(x) |>
+              modify_zero_recode() |>
+              modify_header_rm_md()
+          }
       )
     )
 
@@ -101,7 +100,7 @@ theme_gtsummary_roche <- function(font_size = NULL,
             rlang::expr(gt::opt_table_font(font = "arial")),
             rlang::expr(
               gt::tab_options(
-                table.font.size = 13,
+                table.font.size = !!(font_size %||% 13),
                 data_row.padding = gt::px(1),
                 summary_row.padding = gt::px(1),
                 grand_summary_row.padding = gt::px(1),